requirements.txt
bcrypt==4.2.0
capstone==6.0.0a1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colored-traceback==0.4.2
cryptography==43.0.3
idna==3.10
intervaltree==3.1.0
Mako==1.3.6
MarkupSafe==3.0.2
mpmath==1.3.0
packaging==24.1
paramiko==3.5.0
plumbum==1.9.0
psutil==6.1.0
pwntools==4.13.1
pycparser==2.22
pycryptodome==3.21.0
pyelftools==0.31
Pygments==2.18.0
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
ROPGadget==7.5
rpyc==6.0.1
six==1.16.0
sortedcontainers==2.4.0
sympy==1.13.3
unicorn==2.1.1
unix-ar==0.2.1
urllib3==2.2.3
websockets==13.1
zstandard==0.23.0
